#1406 cleanup and optimizing after refactoring
