+++ processor/src/main/java/org/mapstruct/ap/internal/model/beanmapping/TargetReference.java
@@ -51,12 +51,7 @@ import static org.mapstruct.ap.internal.util.Collections.first;
  *
  * @author Sjaak Derksen
  */
-public class TargetReference {
-
-    private final Parameter parameter;
-    private final List<PropertyEntry> propertyEntries;
-    private final boolean isValid;
-
+public class TargetReference extends AbstractReference {
 
     /**
      * Builds a {@link TargetReference} from an {@code @Mappping}.
@@ -348,84 +343,24 @@ public class TargetReference {
         }
     }
 
-    private TargetReference(Parameter sourceParameter, List<PropertyEntry> sourcePropertyEntries, boolean isValid) {
-        this.parameter = sourceParameter;
-        this.propertyEntries = sourcePropertyEntries;
-        this.isValid = isValid;
-    }
-
-    public Parameter getParameter() {
-        return parameter;
-    }
-
-    public List<PropertyEntry> getPropertyEntries() {
-        return propertyEntries;
-    }
-
-    public boolean isValid() {
-        return isValid;
-    }
-
-    public List<String> getElementNames() {
-        List<String> elementNames = new ArrayList<>();
-        if ( parameter != null ) {
-            // only relevant for source properties
-            elementNames.add( parameter.getName() );
-        }
-        for ( PropertyEntry propertyEntry : propertyEntries ) {
-            elementNames.add( propertyEntry.getName() );
-        }
-        return elementNames;
-    }
-
-    public String getTopPropertyName() {
-        if (  !propertyEntries.isEmpty() ) {
-            return first( propertyEntries ).getFullName();
-        }
-        return null;
+    private TargetReference(Parameter sourceParameter, List<PropertyEntry> targetPropertyEntries, boolean isValid) {
+        super( sourceParameter, targetPropertyEntries, isValid );
     }
 
     public TargetReference pop() {
-        if ( propertyEntries.size() > 1 ) {
-            List<PropertyEntry> newPropertyEntries = new ArrayList<>( propertyEntries.size() - 1 );
-            for ( PropertyEntry propertyEntry : propertyEntries ) {
+        if ( getPropertyEntries().size() > 1 ) {
+            List<PropertyEntry> newPropertyEntries = new ArrayList<>( getPropertyEntries().size() - 1 );
+            for ( PropertyEntry propertyEntry : getPropertyEntries() ) {
                 PropertyEntry newPropertyEntry = propertyEntry.pop();
                 if ( newPropertyEntry != null ) {
                     newPropertyEntries.add( newPropertyEntry );
                 }
             }
-            return new TargetReference( null, newPropertyEntries, isValid );
+            return new TargetReference( null, newPropertyEntries, isValid() );
         }
         else {
             return null;
         }
-    }
-
-    @Override
-    public String toString() {
-
-        String result = "";
-        if ( !isValid ) {
-            result = "invalid";
-        }
-        else if ( propertyEntries.isEmpty() ) {
-            if ( parameter != null ) {
-                result = String.format( "parameter \"%s %s\"", parameter.getType(), parameter.getName() );
-            }
-        }
-        else if ( propertyEntries.size() == 1 ) {
-            PropertyEntry propertyEntry = propertyEntries.get( 0 );
-            result = String.format( "property \"%s %s\"", propertyEntry.getType(), propertyEntry.getName() );
-        }
-        else {
-            PropertyEntry lastPropertyEntry = propertyEntries.get( propertyEntries.size() - 1 );
-            result = String.format(
-                "property \"%s %s\"",
-                lastPropertyEntry.getType(),
-                Strings.join( getElementNames(), "." )
-            );
-        }
-        return result;
     }
 
     private abstract static class MappingErrorMessage {
