+++ processor/src/main/java/org/mapstruct/ap/internal/model/source/Mapping.java
@@ -481,6 +481,9 @@ public class Mapping {
         if ( o == null || getClass() != o.getClass() ) {
             return false;
         }
+        if ( ".".equals( targetName ) ) {
+            return false;
+        }
         Mapping mapping = (Mapping) o;
         return targetName.equals( mapping.targetName );
     }
